:recycle: [Refactor]: FSD 구조 적용 - 6. 슬라이스를 features 레이어로 분리하기
