+++ apps/client/src/features/auth/model/index.ts
@@ -1,0 +1,2 @@
+export { useAuth } from './useAuth';
+export { AuthContext } from './AuthContext';
