+++ apps/client/src/features/chatting/index.ts
@@ -1,0 +1,1 @@
+export { ChatContainer } from './ui';
