+++ apps/client/src/features/editProfile/index.ts
@@ -1,0 +1,1 @@
+export { EditUserInfo } from './ui';
