+++ apps/client/src/features/editProfile/ui/index.ts
@@ -1,0 +1,1 @@
+export { EditUserInfo } from './EditUserInfo';
