+++ apps/client/src/features/watching/index.ts
@@ -1,0 +1,3 @@
+export { useConsumer } from './model';
+
+export { LiveCamperInfo, LivePlayer } from './ui';
