+++ apps/client/src/features/watching/ui/LivePlayer/LivePlayer.tsx
