+++ apps/client/src/pages/Profile/model/index.ts
@@ -1,0 +1,1 @@
+export type { UserData } from './types';
