+++ apps/client/src/pages/Profile/ui/index.ts
@@ -1,4 +1,2 @@
 export { Attendance } from './Attendance';
-export { EditUserInfo } from './EditUserInfo';
 export { UserInfo } from './UserInfo';
-export type { UserData } from './types';
