+++ apps/client/src/shared/api/index.ts
@@ -1,0 +1,2 @@
+export { axiosInstance } from './axios';
+export { useAPI } from './useAPI';
