+++ apps/client/src/shared/contexts/index.ts
@@ -1,2 +1,2 @@
-export { AuthContext } from './AuthContext';
+export { AuthContext } from '../../features/auth/model/AuthContext';
 export { ThemeContext } from './ThemeContext';
