+++ apps/client/src/shared/lib/index.ts
@@ -1,5 +1,3 @@
-export { useAPI } from './useAPI';
-export { useAuth } from './useAuth';
 export { useSocket } from './useSocket';
 export { useTheme } from './useTheme';
 export { useToast } from './useToast';
